chore: enhance tenant isolation and query performance
- Convert tenant_id from INTEGER to BIGINT across all tables
- Add GIN indexes on JSONB columns for faster containment queries
- Add composite indexes for common query patterns (tenant+type, tenant+kb)
- Add check constraints for tenant_id data integrity
- Fix ModelRepository.ClearDefaultByType signature: uint -> uint64 for consistency

+++ internal/application/repository/model.go
@@ -81,7 +81,7 @@ func (r *modelRepository) Delete(ctx context.Context, tenantID uint64, id string
 // This is a batch operation that updates all matching records in one query
 func (r *modelRepository) ClearDefaultByType(
 	ctx context.Context,
-	tenantID uint,
+	tenantID uint64,
 	modelType types.ModelType,
 	excludeID string,
 ) error {
